refactor: init v2
