+++ pkg/apiserver/handler_installer_upgrade.go
@@ -7,7 +7,6 @@ import (
 	appv1alpha1 "bytetrade.io/web3os/app-service/api/app.bytetrade.io/v1alpha1"
 	"bytetrade.io/web3os/app-service/pkg/apiserver/api"
 	"bytetrade.io/web3os/app-service/pkg/constants"
-	"bytetrade.io/web3os/app-service/pkg/helm"
 	"bytetrade.io/web3os/app-service/pkg/kubesphere"
 	"bytetrade.io/web3os/app-service/pkg/utils"
 	apputils "bytetrade.io/web3os/app-service/pkg/utils/app"
@@ -18,31 +17,6 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
-
-func (h *Handler) releaseVersion(req *restful.Request, resp *restful.Response) {
-	appName := req.PathParameter(ParamAppName)
-	owner := req.Attribute(constants.UserContextAttribute)
-	appNamespace, err := utils.AppNamespace(appName, owner.(string), "")
-	if err != nil {
-		api.HandleError(resp, req, err)
-		return
-	}
-
-	actionConfig, _, err := helm.InitConfig(h.kubeConfig, appNamespace)
-	if err != nil {
-		api.HandleError(resp, req, err)
-		return
-	}
-	version, _, err := apputils.GetDeployedReleaseVersion(actionConfig, appName)
-	if err != nil {
-		api.HandleError(resp, req, err)
-		return
-	}
-	resp.WriteAsJson(api.ReleaseVersionResponse{
-		Response: api.Response{Code: 200},
-		Data:     api.ReleaseVersionData{Version: version},
-	})
-}
 
 func (h *Handler) appUpgrade(req *restful.Request, resp *restful.Response) {
 	app := req.PathParameter(ParamAppName)
